Unit Test Completed on HRun and HCapture
