+++ TradeBot/resource/config.properties
@@ -8,13 +8,13 @@ DB_PASS=test123
 APP_PIN=maths
 TRADEBOT_LOG=/resource/logs/tradebot.log
 
-F1HRUN_LOG=/resource/logs/F1HRun.log
-F2HRUN_LOG=/resource/logs/F2HRun.log
-F3HRUN_LOG=/resource/logs/F3HRun.log
+F1_LOG=/resource/logs/F1_Algo.log
+F2_LOG=/resource/logs/F2_Algo.log
+F3_LOG=/resource/logs/F3_Algo.log
 
-F4HCAP_LOG =/resource/logs/F4HCapture.log
-F5HCAP_LOG =/resource/logs/F5HCapture.log
-F6HCAP_LOG =/resource/logs/F6HCapture.log
+F4_LOG =/resource/logs/F4_Algo.log
+F5_LOG =/resource/logs/F5_Algo.log
+F6_LOG =/resource/logs/F6_Algo.log
 
-F7DUMMY = /resource/logs/F6Dummy.log
+F7_LOG = /resource/logs/F6_Algo.log
 TRANSACTION_LOG=/resource/logs/transaction.log
