--- TradeBot/resource/logs/F1HRun.log
@@ -1,72 +1,0 @@
-2018-04-26 14:20:10 [Thread-7] com.tradebot.presto.presto_data_feeder.presto_start_data_feeder()
-INFO: Subscribed - 50496 , NIFTY
-2018-04-26 14:21:32 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: Long buy :Thu Apr 26 14:21:18 IST 2018
-2018-04-26 14:21:38 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: HIGH : 10579.3, LOW : 10575.7
-2018-04-26 14:21:40 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: Trade Switch is OFF, paper order
-2018-04-26 14:23:15 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: long buy and sell Condition3:Thu Apr 26 14:22:48 IST 2018
-2018-04-26 14:23:15 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.UpdateTradeBoard()
-INFO: Selling Price : 10571.4279422875
-2018-04-26 14:23:15 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.UpdateTradeBoard()
-INFO: Buying Price : 10575.976512905
-2018-04-26 14:23:15 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.UpdateTradeBoard()
-INFO: Trade End of the Dau P&L : -4.548570617498626, Percentage % :2.0, Trade Count : 2
-2018-04-26 14:23:16 [Thread-11] com.tradebot.dbcommons.db_commons.getRowCount()
-ERROR: org.h2.jdbc.JdbcSQLException: Syntax error in SQL statement "SELECT * FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='50496' AND TRADESECID='52380' AND ISBUYSELLDONE[*] =='true' "; SQL statement:
-SELECT * FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='50496' and TRADESECID='52380' and ISBUYSELLDONE =='true' [42000-196]
-	at org.h2.engine.SessionRemote.done(SessionRemote.java:629)
-	at org.h2.command.CommandRemote.prepare(CommandRemote.java:83)
-	at org.h2.command.CommandRemote.<init>(CommandRemote.java:49)
-	at org.h2.engine.SessionRemote.prepareCommand(SessionRemote.java:499)
-	at org.h2.jdbc.JdbcConnection.prepareCommand(JdbcConnection.java:1204)
-	at org.h2.jdbc.JdbcStatement.executeQuery(JdbcStatement.java:73)
-	at com.tradebot.dbcommons.db_commons.getRowCount(db_commons.java:66)
-	at com.tradebot.formulas.F1_HRun_Algo.<init>(F1_HRun_Algo.java:60)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:123)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-26 14:23:16 [Thread-11] com.tradebot.dbcommons.db_commons.getRowCount()
-ERROR: org.h2.jdbc.JdbcSQLException: Syntax error in SQL statement "SELECT * FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='50496' AND TRADESECID='52380' AND ISBUYSELLDONE[*] =='true' "; SQL statement:
-SELECT * FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='50496' and TRADESECID='52380' and ISBUYSELLDONE =='true' [42000-196]
-	at org.h2.engine.SessionRemote.done(SessionRemote.java:629)
-	at org.h2.command.CommandRemote.prepare(CommandRemote.java:83)
-	at org.h2.command.CommandRemote.<init>(CommandRemote.java:49)
-	at org.h2.engine.SessionRemote.prepareCommand(SessionRemote.java:499)
-	at org.h2.jdbc.JdbcConnection.prepareCommand(JdbcConnection.java:1204)
-	at org.h2.jdbc.JdbcStatement.executeQuery(JdbcStatement.java:73)
-	at com.tradebot.dbcommons.db_commons.getRowCount(db_commons.java:66)
-	at com.tradebot.formulas.F1_HRun_Algo.<init>(F1_HRun_Algo.java:60)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:123)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-26 14:24:04 [Thread-15] com.tradebot.presto.presto_data_feeder.presto_stop_data_feeder()
-INFO: Unsubscribed - 50496 , NIFTY
-2018-04-26 14:26:51 [Thread-10] com.tradebot.presto.presto_data_feeder.presto_start_data_feeder()
-INFO: Subscribed - 50496 , NIFTY
-2018-04-26 14:28:16 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: Long buy :Thu Apr 26 14:28:19 IST 2018
-2018-04-26 14:28:16 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: HIGH : 10572.95, LOW : 10569.55
-2018-04-26 14:28:16 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: Trade Switch is OFF, paper order
-2018-04-26 14:38:26 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: long buy and sell Condition3:Thu Apr 26 14:38:28 IST 2018
-2018-04-26 14:38:26 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.UpdateTradeBoard()
-INFO: Selling Price : 10569.5
-2018-04-26 14:38:26 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.UpdateTradeBoard()
-INFO: Buying Price : 10572.95
-2018-04-26 14:38:26 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.UpdateTradeBoard()
-INFO: Trade End of the Dau P&L : -3.4500000000007276, Percentage % :2.0, Trade Count : 2
-2018-04-26 14:40:21 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: Long buy :Thu Apr 26 14:40:23 IST 2018
-2018-04-26 14:40:21 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: HIGH : 10572.95, LOW : 10566.8
-2018-04-26 14:40:21 [Thread-11] com.tradebot.formulas.F1_HRun_Algo.CalculateHRun()
-INFO: Trade Switch is OFF, paper order
