--- TradeBot/resource/logs/F2HRun.log
@@ -1,38 +1,0 @@
-2018-04-25 12:25:53 [Thread-11] com.tradebot.formulas.F2_HRun_Algo.assginF1Variables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F2_HRun_Algo.assginF1Variables(F2_HRun_Algo.java:191)
-	at com.tradebot.formulas.F2_HRun_Algo.<init>(F2_HRun_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:124)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:25:53 [Thread-10] com.tradebot.presto.presto_data_feeder.presto_start_data_feeder()
-INFO: Subscribed - 50496 , NIFTY
-2018-04-25 12:25:53 [Thread-11] com.tradebot.formulas.F2_HRun_Algo.assginF1Variables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F2_HRun_Algo.assginF1Variables(F2_HRun_Algo.java:191)
-	at com.tradebot.formulas.F2_HRun_Algo.<init>(F2_HRun_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:124)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:27:37 [Thread-11] com.tradebot.formulas.F2_HRun_Algo.assginF1Variables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F2_HRun_Algo.assginF1Variables(F2_HRun_Algo.java:191)
-	at com.tradebot.formulas.F2_HRun_Algo.<init>(F2_HRun_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:124)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:27:37 [Thread-11] com.tradebot.formulas.F2_HRun_Algo.assginF1Variables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F2_HRun_Algo.assginF1Variables(F2_HRun_Algo.java:191)
-	at com.tradebot.formulas.F2_HRun_Algo.<init>(F2_HRun_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:124)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
