--- TradeBot/resource/logs/F3HRun.log
@@ -1,40 +1,0 @@
-2018-04-25 12:25:53 [Thread-11] com.tradebot.formulas.F3_HRun_Algo.assginF1Variables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F3_HRun_Algo.assginF1Variables(F3_HRun_Algo.java:189)
-	at com.tradebot.formulas.F3_HRun_Algo.<init>(F3_HRun_Algo.java:65)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:125)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:25:53 [Thread-11] com.tradebot.formulas.F3_HRun_Algo.assginF1Variables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F3_HRun_Algo.assginF1Variables(F3_HRun_Algo.java:189)
-	at com.tradebot.formulas.F3_HRun_Algo.<init>(F3_HRun_Algo.java:65)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:125)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:27:37 [Thread-11] com.tradebot.formulas.F3_HRun_Algo.assginF1Variables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F3_HRun_Algo.assginF1Variables(F3_HRun_Algo.java:189)
-	at com.tradebot.formulas.F3_HRun_Algo.<init>(F3_HRun_Algo.java:65)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:125)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:27:37 [Thread-11] com.tradebot.formulas.F3_HRun_Algo.assginF1Variables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F3_HRun_Algo.assginF1Variables(F3_HRun_Algo.java:189)
-	at com.tradebot.formulas.F3_HRun_Algo.<init>(F3_HRun_Algo.java:65)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:125)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:32:55 [Thread-15] com.tradebot.presto.presto_data_feeder.presto_stop_data_feeder()
-INFO: Unsubscribed - 50496 , NIFTY
-2018-04-25 12:32:55 [Thread-15] com.tradebot.presto.presto_data_feeder.presto_stop_data_feeder()
-INFO: Unsubscribed - 50496 , NIFTY
