--- TradeBot/resource/logs/F4HCapture.log
@@ -1,74 +1,0 @@
-2018-04-25 12:25:53 [Thread-11] com.tradebot.formulas.F4_HCapture_Algo.assginFVariables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F4_HCapture_Algo.assginFVariables(F4_HCapture_Algo.java:191)
-	at com.tradebot.formulas.F4_HCapture_Algo.<init>(F4_HCapture_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:126)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:25:53 [Thread-11] com.tradebot.formulas.F5_HCapture_Algo.assginFVariables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F5_HCapture_Algo.assginFVariables(F5_HCapture_Algo.java:190)
-	at com.tradebot.formulas.F5_HCapture_Algo.<init>(F5_HCapture_Algo.java:66)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:127)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:25:54 [Thread-11] com.tradebot.formulas.F6_HCapture_Algo.assginFVariables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F6_HCapture_Algo.assginFVariables(F6_HCapture_Algo.java:191)
-	at com.tradebot.formulas.F6_HCapture_Algo.<init>(F6_HCapture_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:128)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:25:54 [Thread-11] com.tradebot.formulas.F6_HCapture_Algo.assginFVariables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F6_HCapture_Algo.assginFVariables(F6_HCapture_Algo.java:191)
-	at com.tradebot.formulas.F6_HCapture_Algo.<init>(F6_HCapture_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:128)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:27:37 [Thread-10] com.tradebot.presto.presto_data_feeder.presto_start_data_feeder()
-INFO: Subscribed - 50496 , NIFTY
-2018-04-25 12:27:38 [Thread-11] com.tradebot.formulas.F4_HCapture_Algo.assginFVariables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F4_HCapture_Algo.assginFVariables(F4_HCapture_Algo.java:191)
-	at com.tradebot.formulas.F4_HCapture_Algo.<init>(F4_HCapture_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:126)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:27:38 [Thread-11] com.tradebot.formulas.F5_HCapture_Algo.assginFVariables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F5_HCapture_Algo.assginFVariables(F5_HCapture_Algo.java:190)
-	at com.tradebot.formulas.F5_HCapture_Algo.<init>(F5_HCapture_Algo.java:66)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:127)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:27:38 [Thread-11] com.tradebot.formulas.F6_HCapture_Algo.assginFVariables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F6_HCapture_Algo.assginFVariables(F6_HCapture_Algo.java:191)
-	at com.tradebot.formulas.F6_HCapture_Algo.<init>(F6_HCapture_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:128)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
-2018-04-25 12:27:38 [Thread-11] com.tradebot.formulas.F6_HCapture_Algo.assginFVariables()
-ERROR: java.lang.ArrayIndexOutOfBoundsException: 0
-	at com.tradebot.formulas.F6_HCapture_Algo.assginFVariables(F6_HCapture_Algo.java:191)
-	at com.tradebot.formulas.F6_HCapture_Algo.<init>(F6_HCapture_Algo.java:67)
-	at com.tradebot.presto.presto_data_feeder.onFeed(presto_data_feeder.java:128)
-	at com.sft.feedprovider.FeedProvider.update(FeedProvider.java:101)
-	at com.sft.feedprovider.BaseApplication.update(BaseApplication.java:1)
-	at com.sft.feedprovider.Subject.notify(Subject.java:28)
-	at com.sft.feedprovider.DummyMarketDataProvider$MarketPictureFeedConsumer.run(DummyMarketDataProvider.java:114)
