+++ TradeBot/src/com/tradebot/formulas/F2_HCapture_Algo.java
@@ -11,15 +11,16 @@ import com.tradebot.dbcommons.db_commons;
 import com.tradebot.dbcommons.tradebot_utility;
 import com.tradebot.presto.presto_commons;
 
-public class F1_HRun_Algo 
-{
+
+public class F2_HCapture_Algo {
+
 	public double x,y,z,C=0,LC=0,S=0;
 	public Date t1,t2,t3,t4;
 	public int maxtradecount =0, TCount=0, Lcount=0;
 	public boolean isBought =false, isSell = false, isShotsell=false, istradeswitch=false;
 	public double buyPrice,sellPrice,low=0.0,high=0.0,Mpoint,stopl=0.0;
 	public Date fst1, fst2;
-	final static String Fname="F1";
+	final static String Fname="F2";
 	private String configprop=System.getProperty("user.dir")+File.separator+"resource"+File.separator+"config.properties";
 	String tradelogpath;
     tradebot_utility utils = new tradebot_utility();
@@ -36,11 +37,11 @@ public class F1_HRun_Algo
     presto_commons objPresto;
     public Formula_Commons funcom=new Formula_Commons();
     int alreadyEntered = 0;
-    public static String dbtable= "TBL_F1_HRUN_TRADES";
+    public static String dbtable= "TBL_F2_HCAPTURE_TRADES";
 
-	public F1_HRun_Algo(presto_commons objconnect,  String feedid, double tradeprice, int asksize,int bidsize ,String tickdatetime) 
+	public F2_HCapture_Algo(presto_commons objconnect,  String feedid, double tradeprice, int asksize,int bidsize ,String tickdatetime) 
 	{
-		tradelogpath = utils.configlogfile("F1_LOG");
+		tradelogpath = utils.configlogfile("F2_LOG");
 		askvolume = asksize;
 		bidvolume = bidsize;		 
 		objPresto =objconnect;
@@ -50,33 +51,24 @@ public class F1_HRun_Algo
 			 datefmt=new SimpleDateFormat("yyyyMMdd HH:mm:ss");
 			 for(int i =0; i < tradeplayers.size(); i++)
 			 {
-				 if ((dbObj.getRowCount("SELECT * FROM TBL_FORMULA WHERE FEEDSECID ='"+feedid+"' AND TRADESECID ='"+tradeplayers.get(i)+"'") !=0 )&&(dbObj.getSingleCell("SELECT ISEND FROM TBL_FORMULA WHERE FEEDSECID ='"+feedid+"' AND TRADESECID ='"+tradeplayers.get(i)+"' AND FORMULANAME ='"+Fname+"'").equals("false")))
+				 if ((dbObj.getRowCount("SELECT * FROM TBL_FORMULA WHERE FEEDSECID ='"+feedid+"' AND TRADESECID ='"+tradeplayers.get(i)+"'") !=0 )&&(dbObj.getSingleCell("SELECT ISEND FROM TBL_FORMULA WHERE FEEDSECID ='"+feedid+"' AND TRADESECID ='"+tradeplayers.get(i)+"' AND FORMULANAME ='"+Fname+"'").trim().equals("false")))
 				 {
-					 if (dbObj.getRowCount("SELECT * FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"'") == 0)// and ISBUYSELLDONE ='false'
+					 if (dbObj.getRowCount("SELECT * FROM TBL_F2_HCAPTURE_TRADES  WHERE FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"'") == 0)// and ISBUYSELLDONE ='false'
 					 {
-						 dbObj.executeNonQuery("insert into TBL_F1_HRUN_TRADES (FEEDSECID, TRADESECID,ENTRYTIME,BUYPRICE,SELLPRICE,EXITTIME,ISSHOTSELL,HIGH,LOW,ISBOUGHT, ISSELL,MPOINT,EXITCONDITION,TCOUNT,ISBUYSELLDONE,ENTRYID,EXITID,C,LC,S) values "
+						 dbObj.executeNonQuery("insert into TBL_F2_HCAPTURE_TRADES (FEEDSECID, TRADESECID,ENTRYTIME,BUYPRICE,SELLPRICE,EXITTIME,ISSHOTSELL,HIGH,LOW,ISBOUGHT, ISSELL,MPOINT,EXITCONDITION,TCOUNT,ISBUYSELLDONE,ENTRYID,EXITID,C,LC,S) values "
 						 		+ "('"+feedid+"', '"+tradeplayers.get(i)+"', null, 0.0,0.0,null,'false',0.0,0.0,'false','false',0.0,null,0,'false',null,null,0,0,0)");
 					 }
-					 if((dbObj.getRowCount("SELECT * FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"' and ISBUYSELLDONE ='true'") > 0) && (dbObj.getRowCount("SELECT * FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"' and ISBUYSELLDONE ='false'") != 1 ))
+					 if((dbObj.getRowCount("SELECT * FROM TBL_F2_HCAPTURE_TRADES  WHERE FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"' and ISBUYSELLDONE ='true'") > 0) && (dbObj.getRowCount("SELECT * FROM TBL_F2_HCAPTURE_TRADES  WHERE FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"' and ISBUYSELLDONE ='false'") != 1 ))
 					 {
-						String[][] lasttransac =  dbObj.getMultiColumnRecords("SELECT high, low, mpoint, C, LC, S FROM TBL_F1_HRUN_TRADES WHERE id = (SELECT MAX(id) FROM TBL_F1_HRUN_TRADES ) and FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"';");
+						String[][] lasttransac =  dbObj.getMultiColumnRecords("SELECT high, low, mpoint, C, LC, S FROM TBL_F2_HCAPTURE_TRADES WHERE id = (SELECT MAX(id) FROM TBL_F2_HCAPTURE_TRADES ) and FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"';");
 						
-						dbObj.executeNonQuery("insert into TBL_F1_HRUN_TRADES (FEEDSECID, TRADESECID,ENTRYTIME,BUYPRICE,SELLPRICE,EXITTIME,ISSHOTSELL,HIGH,LOW,ISBOUGHT, ISSELL,MPOINT,EXITCONDITION,TCOUNT,ISBUYSELLDONE,ENTRYID,EXITID,C,LC,S) values "
+						dbObj.executeNonQuery("insert into TBL_F2_HCAPTURE_TRADES (FEEDSECID, TRADESECID,ENTRYTIME,BUYPRICE,SELLPRICE,EXITTIME,ISSHOTSELL,HIGH,LOW,ISBOUGHT, ISSELL,MPOINT,EXITCONDITION,TCOUNT,ISBUYSELLDONE,ENTRYID,EXITID,C,LC,S) values "
 						 		+ "('"+feedid+"', '"+tradeplayers.get(i)+"', null, 0.0,0.0,null,'false',"+Double.parseDouble(lasttransac[0][0])+","+Double.parseDouble(lasttransac[0][1])+",'false','false',"+Double.parseDouble(lasttransac[0][2])+",null,0,'false',null,null,"+Integer.parseInt(lasttransac[0][3])+","+Integer.parseInt(lasttransac[0][4])+","+Integer.parseInt(lasttransac[0][5])+")");
 					
 					 }
 					 assginF1Variables(feedid, tradeplayers.get(i),tickdatetime);
+					 CalculateHRun(feedid, tradeplayers.get(i),Double.valueOf(tradeprice), datefmt.parse(tickdatetime));
 					 
-					 //int tradec =Integer.parseInt(dbObj.getSingleCell("SELECT SUM(TCOUNT) as \"count\" FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='"+feedid+"' and TRADESECID='"+tradeplayers.get(i)+"'"));
-					 //TCount = tradec;
-					 //if (tradec < maxtradecount)
-					 //{
-						 CalculateHRun(feedid, tradeplayers.get(i),Double.valueOf(tradeprice), datefmt.parse(tickdatetime));
-					 //}
-					 //else
-					 //{
-					//	 System.out.print("Max trade Count achieved...");
-					 //}
 			 	}
 			 }
 		 }
@@ -112,7 +104,7 @@ public class F1_HRun_Algo
 				stopl=((F1inputdata[0][colstopl] == null ? 0.0 : Double.parseDouble(F1inputdata[0][colstopl])));
 				istradeswitch=((F1inputdata[0][coltradeswitch] == null ? false : Boolean.parseBoolean(F1inputdata[0][coltradeswitch])));
 			 }
-			F1tradedata = dbObj.getMultiColumnRecords("SELECT * FROM TBL_F1_HRUN_TRADES where FEEDSECID='"+feedid+"' and TRADESECID='"+tradeid+"' and ISBUYSELLDONE='false'");
+			F1tradedata = dbObj.getMultiColumnRecords("SELECT * FROM TBL_F2_HCAPTURE_TRADES where FEEDSECID='"+feedid+"' and TRADESECID='"+tradeid+"' and ISBUYSELLDONE='false'");
 			if (F1tradedata != null)
 			{
 				isBought=((F1tradedata[0][tisbought] == null ? false : Boolean.parseBoolean(F1tradedata[0][tisbought])));
@@ -146,23 +138,7 @@ public class F1_HRun_Algo
 			
 			if (ticktime.after(t1))
 			{
-				alreadyEntered = dbObj.getRowCount("SELECT * FROM TBL_F1_HRUN_TRADES WHERE FEEDSECID = '"+feedid+"' and TRADESECID = '"+tradeid+"' and ISBUYSELLDONE ='true'");
-//				if (alreadyEntered == 0)
-//				{
-//					if ((low == 0.0) && (high == 0.0))
-//		    	    {
-//		    	    	low = tickprice;
-//		    	    	high =tickprice;
-//		    	    	dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES SET LOW = "+low+" , HIGH ="+ high +""
-//		    	    			+ " WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
-//		    	    }
-					//if ((isBought==false)&&(isSell==false))
-					//{
-						//if(tickprice > high){high = tickprice;dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET HIGH ="+ high +" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");}
-						//if(tickprice < low){low = tickprice;dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET LOW ="+ low +" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");}   
-					//}
-				//}	
-				
+				alreadyEntered = dbObj.getRowCount("SELECT * FROM TBL_F2_HCAPTURE_TRADES WHERE FEEDSECID = '"+feedid+"' and TRADESECID = '"+tradeid+"' and ISBUYSELLDONE ='true'");
 				if (isBought == true)
 				{
 					// Box 2
@@ -172,27 +148,37 @@ public class F1_HRun_Algo
 						String orderid = null;
 						sellPrice =tickprice;				    	    			
     	    			fst2 = ticktime;
-    	    			Logger.info("long buy and sell Condition3:"+ticktime);
-    	    			TCount =TCount +1;
-    	    			if (istradeswitch ==true)
+    	    			if (C >= 1)
     	    			{
-    	    				if (bidvolume >= 1)
-        	    			{
-    	    					orderid = funcom.LoadDataandOrder(objPresto,feedid, tradeid, "SELL");
-    	    					dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", "
-    	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'true',LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" , EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
-    	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
-    	    				}
+    	    				//Ending Execution for HEAD FEED and trade id
+        	    			dbObj.executeNonQuery("UPDATE TBL_FORMULA  SET ISEND='true' WHERE FEESECID='"+feedid+"' AND TRADESECID='"+tradeid+"' AND FORMULANAME = '"+Fname+"';"); 
+        	    			Logger.info("Head Feed - "+feedid+" & Trade ID - "+tradeid+" && Formula Name - "+Fname+" Achived End point.");
     	    			}
-    	    			else 
+    	    			else
     	    			{
-    	    				dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'true', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    			Logger.info("long buy and sell Condition3:"+ticktime);
+	    	    			TCount =TCount +1;
+	    	    			if (istradeswitch ==true)
+	    	    			{
+	    	    				if (bidvolume >= 1)
+	        	    			{
+	    	    					orderid = funcom.LoadDataandOrder(objPresto,feedid, tradeid, "SELL");
+	    	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", "
+	    	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'true',LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" , EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
+	    	    				}
+	    	    			}
+	    	    			else 
+	    	    			{
+	    	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'true', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    			}
+	    	    			funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, Fname,dbtable);
+	    	    			//Ending Execution for HEAD FEED
+	    	    			dbObj.executeNonQuery("UPDATE TBL_FORMULA  SET ISEND='true' WHERE FEEDSECID='"+feedid+"' AND TRADESECID='"+tradeid+"' AND FORMULANAME = '"+Fname+"';"); 
+	    	    			Logger.info("Head Feed - "+feedid+" & Trade ID - "+tradeid+" && Formula Name - "+Fname+" Achived End point.");
     	    			}
-    	    			funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, "F1",dbtable);
-    	    			//Ending Execution for HEAD FEED
-    	    			dbObj.executeNonQuery("UPDATE TBL_FORMULA  SET ISEND='true' WHERE FEEDSECID='"+feedid+"' AND TRADESECID='"+tradeid+"' AND FORMULANAME = '"+Fname+"';"); 
-    	    			Logger.info("Head Feed - "+feedid+" & Trade ID - "+tradeid+" && Formula Name - "+Fname+" Achived End point.");
-					}
+    	    			
+    	    		}
 					// Box 3
 					else if(tickprice < Mpoint - (Mpoint*(y/100)))
 					{
@@ -200,7 +186,28 @@ public class F1_HRun_Algo
 						sellPrice =tickprice;				    	    			
     	    			fst2 = ticktime;
     	    			Logger.info("long buy and sell Condition3:"+ticktime);
-    	    			TCount =TCount +1;
+    	    			//Boolean isupdated=false;
+    	    			if (C == 0)
+    	    			{
+    	    				TCount =TCount +1;
+    	    				//isupdated = true;
+    	    				if (istradeswitch ==true)
+        	    			{
+        	    				if (bidvolume >= 1)
+            	    			{
+        	    					orderid = funcom.LoadDataandOrder(objPresto,feedid, tradeid, "SELL");
+        	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", "
+        	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'true' , LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+",  EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+        	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
+        	    				}
+        	    			}
+        	    			else 
+        	    			{
+        	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'true', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+        	    			}
+        	    			funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, Fname,dbtable);
+    	    			}
+    	    			
     	    			Mpoint = Mpoint - (Mpoint*(y/100));
     	    			C = C - 1;
     	    			if (C == -1)
@@ -229,29 +236,42 @@ public class F1_HRun_Algo
     	    			{
     	    				C=0;
     	    			}
-    	    			if (istradeswitch ==true)
+    	    			
+    	    			dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET Tcount="+TCount+", LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and id = (SELECT MAX(id) FROM TBL_F2_HCAPTURE_TRADES )");
+    	    			if (dbObj.getRowCount("SELECT * FROM TBL_F2_HCAPTURE_TRADES WHERE TCOUNT = 2 and ISBUYSELLDONE = 'false'") == 1)
     	    			{
-    	    				if (bidvolume >= 1)
-        	    			{
-    	    					orderid = funcom.LoadDataandOrder(objPresto,feedid, tradeid, "SELL");
-    	    					dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", "
-    	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'true' , LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+",  EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
-    	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
-    	    				}
+    	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ISBUYSELLDONE = 'true' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and id = (SELECT MAX(id) FROM TBL_F2_HCAPTURE_TRADES )"); 	    			
     	    			}
-    	    			else 
-    	    			{
-    	    				dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'true', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
-    	    			}
-    	    			funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, "F1",dbtable);
     	    			//calculatefigure(feedid,tradeid);	
 					}
 					// Box 4
 					else if(tickprice >= Mpoint + (Mpoint*(z/100)))
 					{
+						String orderid = null;
 						Mpoint= Mpoint + (Mpoint*(z/100));
+						sellPrice = tickprice;
 						C=C+1;
-						dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES SET LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+						//dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES SET LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+    	    			if (C == 1)
+    	    			{
+    	    				TCount = TCount +1 ;
+    	    				if (istradeswitch ==true)
+	    	    			{
+	    	    				if (bidvolume >= 1)
+	        	    			{
+	    	    					orderid = funcom.LoadDataandOrder(objPresto,feedid, tradeid, "SELL");
+	    	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", "
+	    	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'false' , LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+",  EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
+	    	    				}
+	    	    			}
+	    	    			else 
+	    	    			{
+	    	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',SELLPRICE ="+sellPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'false', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    			}
+    	    				funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, Fname,dbtable);
+    	    			}
+    	    			dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET Tcount="+TCount+", LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and id = (SELECT MAX(id) FROM TBL_F2_HCAPTURE_TRADES )");
     	    			
 					}
 					
@@ -260,9 +280,32 @@ public class F1_HRun_Algo
 	    	    {
 					if (tickprice <= (Mpoint - (Mpoint*(z/100))))
 					{
+						String orderid = null;
 						Mpoint = Mpoint - (Mpoint*(z/100));
 						C=C+1;
-						dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES SET LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+						//dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES SET LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+						buyPrice =tickprice;
+						if (C == 1)
+						{
+							TCount = TCount +1 ;
+							if (istradeswitch ==true)
+	    	    			{
+	    	    				if (askvolume >= 1)
+	        	    			{
+	    	    					orderid = funcom.LoadDataandOrder(objPresto,feedid, tradeid, "BUY");
+	    	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", "
+	    	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'false' , LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+",  EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
+	    	    				}
+	    	    			}
+	    	    			else 
+	    	    			{
+	    	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'false', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    			}
+							funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, Fname,dbtable);
+						}
+						dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET Tcount="+TCount+", LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and id = (SELECT MAX(id) FROM TBL_F2_HCAPTURE_TRADES )");
+    	    			
 					}
 					else if(tickprice > (Mpoint + (Mpoint*(y/100))))
 					{
@@ -270,7 +313,26 @@ public class F1_HRun_Algo
 						buyPrice =tickprice;				    	    			
     	    			fst2 = ticktime;
     	    			Logger.info("short sell buy Condition2:"+ticktime);
-    	    			TCount =TCount +1;
+
+    	    			if (C==0)
+    	    			{
+    	    				TCount =TCount +1;
+    	    				if (istradeswitch ==true)
+        	    			{
+        	    				if (askvolume >= 1)
+            	    			{
+        	    					orderid = funcom.LoadDataandOrder(objPresto, feedid, tradeid, "BUY");
+        	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", "
+        	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'true' , LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+",  EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+        	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
+        	    				}
+        	    			}
+        	    			else 
+        	    			{
+        	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'true', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+        	    			}
+        	    			funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, Fname,dbtable);
+    	    			}
     	    			Mpoint = Mpoint + (Mpoint*(y/100));
     	    			C=C-1;
     	    			if (C == -1)
@@ -297,22 +359,12 @@ public class F1_HRun_Algo
     	    			{
     	    				C =0;
     	    			}
+    	    			dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET Tcount="+TCount+", LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and id = (SELECT MAX(id) FROM TBL_F2_HCAPTURE_TRADES )"); 	    			
+    	    			if (dbObj.getRowCount("SELECT * FROM TBL_F2_HCAPTURE_TRADES WHERE TCOUNT = 2 and ISBUYSELLDONE = 'false'") == 1)
+    	    			{
+    	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ISBUYSELLDONE = 'true' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and id = (SELECT MAX(id) FROM TBL_F2_HCAPTURE_TRADES )"); 	    			
+    	    			}
     	    			
-    	    			if (istradeswitch ==true)
-    	    			{
-    	    				if (askvolume >= 1)
-        	    			{
-    	    					orderid = funcom.LoadDataandOrder(objPresto, feedid, tradeid, "BUY");
-    	    					dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", "
-    	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'true' , LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+",  EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
-    	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
-    	    				}
-    	    			}
-    	    			else 
-    	    			{
-    	    				dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'true', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
-    	    			}
-    	    			funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, "F1",dbtable);
     	    			//calculatefigure(feedid,tradeid);
 					}
 					else if(ticktime.after(t4))
@@ -322,25 +374,37 @@ public class F1_HRun_Algo
     	    			fst2 = ticktime;
     	    			Logger.info("short buy Condition3:"+ticktime);
     	    			TCount =TCount +1;
-    	    			if (istradeswitch ==true)
+    	    			if (C >= 1)
     	    			{
-    	    				if (askvolume >= 1)
-        	    			{
-    	    					orderid = funcom.LoadDataandOrder(objPresto, feedid, tradeid, "BUY");
-    	    					dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", "
-    	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'true',LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" , EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
-    	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
-    	    				}
+    	    				//Ending Execution for HEAD FEED
+	    	    			dbObj.executeNonQuery("UPDATE TBL_FORMULA  SET ISEND='true' WHERE FEESECID='"+feedid+"' AND TRADESECID='"+tradeid+"' AND FORMULANAME = '"+Fname+"';"); 
+	    	    			Logger.info("Head Feed - "+feedid+" & Trade ID - "+tradeid+" && Formula Name - "+Fname+" Achived End point.");
+    	    			
     	    			}
-    	    			else 
+    	    			else
     	    			{
-    	    				dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'true', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+"  WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    			if (istradeswitch ==true)
+	    	    			{
+	    	    				if (askvolume >= 1)
+	        	    			{
+	    	    					orderid = funcom.LoadDataandOrder(objPresto, feedid, tradeid, "BUY");
+	    	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", "
+	    	    							+ "Tcount="+TCount+", ISBUYSELLDONE = 'true',LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" , EXITID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
+	    	    				}
+	    	    			}
+	    	    			else 
+	    	    			{
+	    	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET EXITCONDITION='N/A',EXITTIME='"+ticktime.toString()+"',BUYPRICE ="+buyPrice+", Tcount="+TCount+", ISBUYSELLDONE = 'true', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+"  WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
+	    	    			}
+	    	    			funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, Fname,dbtable);
+	    	    			//calculatefigure(feedid,tradeid);
+	    	    			//Ending Execution for HEAD FEED
+	    	    			dbObj.executeNonQuery("UPDATE TBL_FORMULA  SET ISEND='true' WHERE FEEDSECID='"+feedid+"' AND TRADESECID='"+tradeid+"' AND FORMULANAME = '"+Fname+"';"); 
+	    	    			Logger.info("Head Feed - "+feedid+" & Trade ID - "+tradeid+" && Formula Name - "+Fname+" Achived End point.");
     	    			}
-    	    			funcom.calculatefigure(sellPrice, buyPrice, feedid,tradeid, "F1",dbtable);
-    	    			//calculatefigure(feedid,tradeid);
-    	    			//Ending Execution for HEAD FEED
-    	    			dbObj.executeNonQuery("UPDATE TBL_FORMULA  SET ISEND='true' WHERE FEEDSECID='"+feedid+"' AND TRADESECID='"+tradeid+"' AND FORMULANAME = '"+Fname+"';"); 
-    	    			Logger.info("Head Feed - "+feedid+" & Trade ID - "+tradeid+" && Formula Name - "+Fname+" Achived End point.");
+    	    			dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET Tcount="+TCount+", LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and id = (SELECT MAX(id) FROM TBL_F2_HCAPTURE_TRADES )");
+    	    			
 					}
 	    	    }
 				// branching Box 1  buyleg or shortleg box 1
@@ -350,11 +414,11 @@ public class F1_HRun_Algo
 		    	    {
 		    	    	low = tickprice;
 		    	    	high =tickprice;
-		    	    	dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES SET LOW = "+low+" , HIGH ="+ high +""
+		    	    	dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES SET LOW = "+low+" , HIGH ="+ high +""
 		    	    			+ " WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
 		    	    }
-					if(tickprice > high){high = tickprice;dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET HIGH ="+ high +" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");}
-					if(tickprice < low){low = tickprice;dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET LOW ="+ low +" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");} 
+					if(tickprice > high){high = tickprice;dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET HIGH ="+ high +" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");}
+					if(tickprice < low){low = tickprice;dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET LOW ="+ low +" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");} 
 					double buylegprice = low + (low*(x/100)), sellingprice = high - (high*(x/100));
 					// buy leg box 1
 					if (tickprice > buylegprice)
@@ -383,7 +447,7 @@ public class F1_HRun_Algo
         	    				if (askvolume >= 1)
 	        	    			{
         	    					orderid = funcom.LoadDataandOrder(objPresto, feedid, tradeid, "BUY");
-        	    					dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET ENTRYTIME = '"+ticktime.toString()+"',"
+        	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ENTRYTIME = '"+ticktime.toString()+"',"
         	    							+ "BUYPRICE="+buyPrice+",ISSHOTSELL='"+isShotsell+"',  isBought='"+isBought+"', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" ,Tcount="+TCount+", ENTRYID='"+orderid+"' WHERE FEEDSECID='"+feedid+"'"
         	    									+ " and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");   
         	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
@@ -392,7 +456,7 @@ public class F1_HRun_Algo
         	    			else
         	    			{
         	    				Logger.info("Trade Switch is OFF, paper order");
-        	    				dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET ENTRYTIME = '"+ticktime.toString()+"',"
+        	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ENTRYTIME = '"+ticktime.toString()+"',"
         	    						+"BUYPRICE="+buyPrice+",ISSHOTSELL='"+isShotsell+"',  isBought='"+isBought+"', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" ,Tcount="+TCount+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
         	    			}
         	    		}
@@ -424,7 +488,7 @@ public class F1_HRun_Algo
 	        	    				if (bidvolume >= 1)
 		        	    			{
 	        	    					orderid = funcom.LoadDataandOrder(objPresto, feedid, tradeid, "SELL");
-	        	    					dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET ISSHOTSELL='"+isShotsell+"', ENTRYTIME='"+ticktime.toString()+"',SELLPRICE="+sellPrice+","
+	        	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ISSHOTSELL='"+isShotsell+"', ENTRYTIME='"+ticktime.toString()+"',SELLPRICE="+sellPrice+","
 	    	        	    					+ " isSell='"+isSell+"',Tcount="+TCount+",LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" , ENTRYID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' "
 	    	        	    							+ "and ISBUYSELLDONE ='false'");   
 	        	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
@@ -432,7 +496,7 @@ public class F1_HRun_Algo
 	        	    			}
 	        	    			else
 	        	    			{
-	        	    			dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET ISSHOTSELL='"+isShotsell+"', ENTRYTIME='"+ticktime.toString()+"',SELLPRICE="+sellPrice+","
+	        	    			dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ISSHOTSELL='"+isShotsell+"', ENTRYTIME='"+ticktime.toString()+"',SELLPRICE="+sellPrice+","
 	        	    					+ " isSell='"+isSell+"',Tcount="+TCount+", LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' "
 	        	    							+ "and ISBUYSELLDONE ='false'");
 	        	    			}
@@ -470,7 +534,7 @@ public class F1_HRun_Algo
         	    				if (askvolume >= 1)
 	        	    			{
         	    					orderid = funcom.LoadDataandOrder(objPresto, feedid, tradeid, "BUY");
-        	    					dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET ENTRYTIME = '"+ticktime.toString()+"',"
+        	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ENTRYTIME = '"+ticktime.toString()+"',"
         	    							+ "BUYPRICE="+buyPrice+",ISSHOTSELL='"+isShotsell+"',  isBought='"+isBought+"', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" ,Tcount="+TCount+", ENTRYID='"+orderid+"' WHERE FEEDSECID='"+feedid+"'"
         	    									+ " and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");   
         	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
@@ -479,7 +543,7 @@ public class F1_HRun_Algo
         	    			else
         	    			{
         	    				Logger.info("Trade Switch is OFF, paper order");
-        	    				dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET ENTRYTIME = '"+ticktime.toString()+"',"
+        	    				dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ENTRYTIME = '"+ticktime.toString()+"',"
         	    						+"BUYPRICE="+buyPrice+",ISSHOTSELL='"+isShotsell+"',  isBought='"+isBought+"', LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" ,Tcount="+TCount+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='false'");
         	    			}
         	    		}
@@ -509,7 +573,7 @@ public class F1_HRun_Algo
 	        	    				if (bidvolume >= 1)
 		        	    			{
 	        	    					orderid = funcom.LoadDataandOrder(objPresto, feedid, tradeid, "SELL");
-	        	    					dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET ISSHOTSELL='"+isShotsell+"', ENTRYTIME='"+ticktime.toString()+"',SELLPRICE="+sellPrice+","
+	        	    					dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ISSHOTSELL='"+isShotsell+"', ENTRYTIME='"+ticktime.toString()+"',SELLPRICE="+sellPrice+","
 	    	        	    					+ " isSell='"+isSell+"',Tcount="+TCount+",LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" , ENTRYID='"+orderid+"' WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' "
 	    	        	    							+ "and ISBUYSELLDONE ='false'");   
 	        	    					Logger.info("Trade Switch is ON, Order Palced : "+orderid);
@@ -517,7 +581,7 @@ public class F1_HRun_Algo
 	        	    			}
 	        	    			else
 	        	    			{
-	        	    			dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET ISSHOTSELL='"+isShotsell+"', ENTRYTIME='"+ticktime.toString()+"',SELLPRICE="+sellPrice+","
+	        	    			dbObj.executeNonQuery("UPDATE TBL_F2_HCAPTURE_TRADES  SET ISSHOTSELL='"+isShotsell+"', ENTRYTIME='"+ticktime.toString()+"',SELLPRICE="+sellPrice+","
 	        	    					+ " isSell='"+isSell+"',Tcount="+TCount+", LOW ="+low+", HIGH="+high+", Mpoint="+Mpoint+",C="+C+", LC="+LC+", S="+S+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' "
 	        	    							+ "and ISBUYSELLDONE ='false'");
 	        	    			}
@@ -537,153 +601,5 @@ public class F1_HRun_Algo
 			
 		}
 	}
-//	public double CalculateTax(double price)
-//	{
-//		double finalcost =0.0, Brokerage=0.0,SST=0.0,TC=0.0,ST=0.0,SEBI=0.0,STAMP=0.0;
-//		try
-//		{
-//			Brokerage = (price*0.006)/100;
-//			SST=(price*0.0125)/100;
-//			TC=(price*0.00325)/100;
-//			ST = (Brokerage+TC)*18/100;
-//			SEBI=(price*0.002)/100;
-//			STAMP=(price*0.006)/100;
-//			finalcost = price - (Brokerage + SST + TC + ST + SEBI + STAMP);
-//		}
-//		catch(Exception ex)
-//		{
-//			Logger.error(ex);
-//		}
-//		return finalcost;
-//	}
-//	
-//	public void calculatefigure(String feedid, String tradeid)
-//	{
-//		
-//		try
-//		{
-//			sellPrice = CalculateTax(sellPrice);
-//			buyPrice = CalculateTax(buyPrice);
-//			//dbObj.executeNonQuery("UPDATE TBL_F1_HRUN_TRADES  SET SELLPRICE="+sellPrice+", BUYPRICE ="+buyPrice+" WHERE FEEDSECID='"+feedid+"' and TRADESECID ='"+tradeid+"' and ISBUYSELLDONE ='true'");
-//			UpdateTradeBoard(feedid, tradeid);
-//		}
-//		catch(Exception ex)
-//		{
-//			Logger.error(ex);
-//		}
-//	}
-//	public void UpdateTradeBoard(String feedid, String tradeid)
-//	{
-//		String [][] pldata;
-//		double fpl=0.00, fpercent=0;
-//		DecimalFormat f = new DecimalFormat("##.00");
-//		try
-//		{
-//			pldata = dbObj.getMultiColumnRecords("SELECT SUM(SELLPRICE) as \"TOTAL_SELL\",SUM(BUYPRICE) as \"TOTAL_BUY\",  SUM(TCOUNT) as \"TOTAL_TRADE\" FROM TBL_F1_HRUN_TRADES  WHERE FEEDSECID='"+feedid+"' and TRADESECID='"+tradeid+"' and ISBUYSELLDONE='true';");
-//			if (pldata != null)
-//			{
-//				fpl = Double.parseDouble(pldata[0][0]) - Double.parseDouble(pldata[0][1]);
-//				double avg = (fpl*100) / (Integer.parseInt(pldata[0][2]));
-//				fpercent = (fpl*100)/avg;
-//				Logger.info("Selling Price : "+ pldata[0][0] );
-//	        	Logger.info("Buying Price : "+pldata[0][1] );
-//	        	Logger.info("Trade End of the Dau P&L : "+fpl +", Percentage % :"+fpercent +", Trade Count : "+pldata[0][2]);
-//	        	dbObj.executeNonQuery("UPDATE  TBL_TRADEBOARD SET F1PC="+f.format(fpercent)+", F1TC="+Integer.parseInt(pldata[0][2])+", F1PL="+ f.format(fpl)+" WHERE FEEDSECID ='"+feedid+"' and TRADESECID = '"+tradeid+"'");
-//			}
-//		}
-//		catch(Exception ex)
-//		{
-//			Logger.error(ex);
-//		}
-//	}
-//	public String getseccode(String sectype)
-//	{
-//		String secCode=null;
-//		try
-//		{
-//			switch (sectype) {
-//			case "Equities":
-//				secCode = "CM";
-//				break;
-//			case "FUTIDX":
-//			case "FUTSTK":
-//			case "OPTIDX":
-//			case "OPTSTK":
-//			case "FUTCOM":
-//				secCode = "FO";
-//				break;
-//
-//			default:
-//				secCode = "CM";
-//				break;
-//			}
-//		}
-//		catch(Exception ex)
-//		{
-//		   	
-//		}
-//		return secCode;
-//	}
-//	public String LoadDataandOrder(String feedid, String tradeid, String order)
-//	{
-//		String OrderId = null;
-//		try
-//		{
-//			//objPresto.PlaceOrder("omnesys","CM",tradeid,String.valueOf(tradeid) ,"22-02-2018","FA9749","1","0.0","0.0","na","0.0","MARKET",
-//			//		"Presto_Mathsartz_Strategy","TestOrders","DAY","BUY");
-//			String ESB_EXCHANGE, SECURITY_TYPE, ESB_SYMBOL, ESB_SECURITYID,
-//			 EXP_DATE, ESB_ACCOUNT, QUAN_TITY, T_PRICE, STOP_PRICE,	
-//			 ESB_OPTIONTYPE, ESB_STRIKEPPRICE, ORDER_TYPE, INSTANCEID_CUSTOMFIELD, T_REMARK,
-//			 TIME_INFORCE, T_SIDE;
-//			
-//			ESB_EXCHANGE = "omnesys";
-//			String [][] playerdet = dbObj.getMultiColumnRecords("SELECT FEEDSECID ,TRADESECID ,SYMBOL ,INSTTYPE ,EXPIRYDD ,EXPIRYMMMYY ,OPTTYPE ,STRIKEPRICE FROM TBL_TRADERS WHERE TRADESECID='"+tradeid+"' AND FEEDSECID ='"+feedid+"' ;");
-//			SECURITY_TYPE = getseccode(playerdet[0][3].toString().trim());
-//			ESB_SYMBOL = playerdet[0][2].toString().trim();
-//			ESB_SECURITYID = tradeid;
-//			
-//			if (playerdet[0][4].toString().trim().equals("0"))
-//			{
-//				SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
-//				 Date date = new Date();
-//				String formattedDate = formatter.format(date);
-//				EXP_DATE = formattedDate;
-//			}
-//			else
-//			{
-//				SimpleDateFormat format1 = new SimpleDateFormat("ddMMMyy");
-//			    SimpleDateFormat format2 = new SimpleDateFormat("dd-MM-yy");
-//			    Date date = format1.parse(playerdet[0][4].trim()+playerdet[0][5].trim());
-//			    System.out.println(format2.format(date));
-//				EXP_DATE = format2.format(date).toString();
-//			}
-//			ESB_ACCOUNT = "FA9749";
-//			QUAN_TITY = dbObj.getSingleCell("SELECT LOTSIZE FROM TBL_FORMULA WHERE TRADESECID = '"+tradeid+"' AND FEEDSECID= '"+feedid+"'");
-//			T_PRICE="0.0";
-//			STOP_PRICE="0.0";
-//			ESB_OPTIONTYPE=playerdet[0][6].toString().trim();
-//			ESB_STRIKEPPRICE =playerdet[0][7].toString().trim();
-//			ORDER_TYPE="MARKET";
-//			INSTANCEID_CUSTOMFIELD="Presto_Mathsartz_Strategy";
-//			T_REMARK="Test";
-//			TIME_INFORCE="DAY"; 
-//			T_SIDE=order.toUpperCase();
-//			try 
-//			{
-//				OrderId =  objPresto.PlaceOrder(ESB_EXCHANGE, SECURITY_TYPE, ESB_SYMBOL, ESB_SECURITYID,
-//						 EXP_DATE, ESB_ACCOUNT, QUAN_TITY, T_PRICE, STOP_PRICE,	
-//						 ESB_OPTIONTYPE, ESB_STRIKEPPRICE, ORDER_TYPE, INSTANCEID_CUSTOMFIELD, T_REMARK,
-//						 TIME_INFORCE, T_SIDE);
-//			}
-//			catch (Exception e) {
-//				// TODO: handle exception
-//			}
-//			
-//		}
-//		catch(Exception ex)
-//		{
-//			Logger.error(ex);
-//		}
-//		return OrderId;
-//	}
+
 }
